+++ R/Heterozygosity.R
@@ -1,11 +1,13 @@
 #' A function to estimate heterozygosity.
 #'
-#' @param VCF Character string indicating the name of the vcf file to be used in analysis.
-#' @param pops Character string indicating the name of the population assignment file. This file should have four columns and be in the same order as your vcf file. The first column named Sample indicates the sample name. The second column named Population indicates the population assignment of each individual. The third column named Longitude indicates the longitude of the sample.  The fourth column named Latitude indicates the latitude of the sample.
-#' @param ploidy Numeric. The ploidy of the data.
-#' @param prefix Character string that will be appended to file output.
+#' @param data Character. String indicating the name of the vcf file or vcfR object to be used in the analysis.
+#' @param pops Character. String indicating the name of the population assignment file or dataframe containing the population assignment information for each individual in the data. This file must be in the same order as the vcf file and include columns specifying the individual and the population that individual belongs to. The first column should contain individual names and the second column should indicate the population assignment of each individual. Alternatively, you can indicate the column containing the individual and population information using the individual_col and population_col arguments.
+#' @param statistic Character. String indicating the statistic to calculate. Options are ...
 #' @param write Boolean. Whether or not to write the output to a file in the current working directory.
-#'
+#' @param prefix Character. String that will be appended to file output.
+#' @param population_col Numeric. Optional argument (a number) indicating the column that contains the population assignment information.
+#' @param individual_col Numeric. Optional argument (a number) indicating the column that contains the individuals (i.e., sample name) in the data.
+
 #' @return A list containing the estimated diversity statistics, model output from linear regression of these statistics against latitude, and model plots.
 #' @export
 #'
@@ -15,7 +17,7 @@
 #' data("HornedLizard_VCF")
 #' Test <- Div_stats(VCF = HornedLizard_VCF, pops = HornedLizard_Pop,
 #' ploidy = 2, write = FALSE)}
-Heterozygosity <- function(data, pops, ploidy, statistic, write = FALSE, prefix, population_col = NULL, individual_col = NULL) {
+Heterozygosity <- function(data, pops, statistic, write = FALSE, prefix, population_col = NULL, individual_col = NULL) {
   Latitude <- Heterozygosity <- Pop <- Standard.Deviation <- NULL
 
   # Read in population assignment data
@@ -64,7 +66,7 @@ Heterozygosity <- function(data, pops, ploidy, statistic, write = FALSE, prefix,
     Dat <- sapply(Dat, as.numeric)
   }
   else if(tools::file_ext(data) == 'geno'){
-    Dat <- utils::read.table(Data)
+    Dat <- utils::read.table(data)
     print("Geno file detected, proceeding to formatting. Note, PopGenHelpR assumes that your individuals in the geno file and
           popmap are in the same order, please check to avoid erroneous inferences.")
   }
@@ -126,84 +128,13 @@ Heterozygosity <- function(data, pops, ploidy, statistic, write = FALSE, prefix,
  ExpHet_res_perloc  <- lapply(Dat_perpop, ExpHe)
  ExpHet_res_avg  <- lapply(ExpHet_res_perloc, mean)
 
- PropHt <-
- StHe <-
- StHo <-
- IR <-
- HL <-
+ PropHt <- NULL
+ StHe <- NULL
+ StHo <- NULL
+ IR <- NULL
+ HL <- NULL
 
+ Output <- list(Obs_Het_res_avg, ObsHet_res_perloc)
 
-
-
-
-
-
-
-  ##########################
-  ##### Heterozygosity #####
-  ##########################
-  # Calculate heterozygosity and standard deviation for each population
-  Het <- hierfstat::basic.stats(Hstat)
-
-  # Get per populations heterozygosity
-  H_all <- data.frame(colMeans(Het$Ho, na.rm = TRUE))
-  H_all$Pop <- rownames(H_all)
-  colnames(H_all) <- c('Heterozygosity', 'Pop')
-
-  # Standard deviation
-  H_all$StandardDeviation <- stats::sd(H_all$Heterozygosity)
-
-  # Attach coordinates
-  # First we check to make sure that the populations are in the right order
-  if(any(H_all$Pop != unique(P[,2]))){
-    stop("Populations are not in the correct order, if you see this please email the package authors")
-  }
-
-  # Pull populations coordinates
-  Pop_coords <- P[!duplicated(P$Population),]
-
-  # Append to heterozygosity estimates
-  H_all[,4:5] <- Pop_coords[,3:4]
-  colnames(H_all) <- c('Heterozygosity', 'Pop', 'Standard.Deviation','Longitude', 'Latitude')
-
-  # Is there a statistical relationship between heterozygosity and latitude
-  Het_model <- stats::lm(H_all$Heterozygosity ~ H_all$Latitude)
-
-  message('Heterozygosity calculated, moving to private alleles')
-
-
-  ##########################
-  ##### Visualizations #####
-  ##########################
-  ##### Heterozygosity #####
-  # We will make a plot of heterozygostiy estimates and an interpolated map of heterozygosity values for each population
-
-  # Plot of heterozygosity values (y-axis) and latitude (x-axis)
-  Het_plot <- ggplot2::ggplot(data = H_all, ggplot2::aes(x = Latitude, y = Heterozygosity)) + ggplot2::geom_point(ggplot2::aes(color = Pop),size = 3) +
-    ggplot2::geom_errorbar(data = H_all, ggplot2::aes(x = Latitude, ymin = Heterozygosity-Standard.Deviation, ymax = Heterozygosity+Standard.Deviation,
-                                                      color = Pop)) +
-    ggplot2::stat_smooth(method = 'lm', formula = y ~ x, size =1, colour = 'black') +
-    ggplot2::labs(x = 'Latitude', y = 'Heterozygosity') +
-    ggplot2::ggtitle(expression(atop("Obersved Heterozygosity of Localities"))) +
-    ggplot2::theme_classic() + ggplot2::theme(plot.title = ggplot2::element_text(hjust = 0.5),
-                                              legend.title = ggplot2::element_blank())
-
-
-  Output <- list(H_all, Het_model, Het_plot)
-  names(Output) <- c('Heterozygosity_calculations', "Heterozygosity_vs_Latitude_Model",
-                     "Heterozygosity_vs_Latitude_plot")
-  if(write == TRUE){
-    # Write out heterozygosity results
-    utils::write.csv(H_all, file = paste(as.character(prefix), '_Heterozygosity.csv', sep = ''), row.names = FALSE)
-    summary(Het_model)
-    # Write out linear regression results heterozygosity ~ latitude
-    sink(paste(as.character(prefix), '_Heterozygosity_lm.txt', sep = ''))
-    summary(Het_model)
-    sink()
-  }
-
-  message("Calculations have finished, the packages used to perform file formatting and calculations were
-  vcfR, adegenet, and dartR for formatting, hierfstat to calculate heterozygosity, and poppr to calculate private alleles")
-
-  return(Output)
+return(Output)
 }